Add card state methods and transaction validations

+++ Contracts/ICardService.cs
@@ -4,7 +4,7 @@ namespace w13_Quiz.Contracts
 {
     public interface ICardService
     {
-        void CreateCard(string cardNumber,  string holderName,  string password, float balance);
+        void CreateCard(string cardNumber,  string holderName,  string password, decimal balance);
 
         IEnumerable<ReceivedTransactionDto> GetCardReceivedTransactions(string cardNumber);
         IEnumerable<SentTransactionDto> GetCardSentTransactions(string cardNumber);
